+++ amundsen_application/static/js/components/Card/styles.scss
